Testing and coverage refactor
